requirements.txt
absl-py==2.3.1
asttokens==3.0.0
async-timeout==5.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
cv==1.0.0
debugpy==1.8.16
decorator==5.2.1
etils==1.13.0
exceptiongroup==1.3.0
executing==2.2.0
fsspec==2025.7.0
glfw==2.9.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mujoco==3.3.4
nest-asyncio==1.6.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
PyOpenGL==3.1.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.2
redis==6.2.0
requests==2.32.4
rich==14.1.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
zipp==3.23.0
